Sveskesuppe

+++ src/Part1/WindowHandler.java
@@ -19,13 +19,13 @@ public class WindowHandler {
 			
 	// Converts X-pixel to coordinate
 	private static double pixelToGeoX(int x) {
-		return  ((x/ Window.use().getMapWidth()) * geoWidth);
+		return  (((double)x/ Window.use().getMapWidth()) * geoWidth);
 	}
 	
 	
 	// Converts Y-pixel to coordinate
 	private static double pixelToGeoY(int y) {
-		return  ((y/(Window.use().getMapHeight()) * geoHeight));
+		return  (((double)y/(Window.use().getMapHeight()) * geoHeight));
 	}
 	
 	// Takes a search area in pixels, and saves a list of all the edges to be drawn on the map
